Extract fetching process
